// projects/zweihander-1e/enums.ts
export enum DrawbackId {
  BAD_TICKER = 'drawback:bad_ticker',
  BLACK_CATARACT = 'drawback:black_cataract',
  BLEEDER = 'drawback:bleeder',
  BRANDED = 'drawback:branded',
  CHOLERIC_TEMPERAMENT = 'drawback:choleric_temperament',
  CROP_EAR = 'drawback:crop_ear',
  CURSED = 'drawback:cursed',
  DEAL_WITH_THE_DEVIL = 'drawback:deal_with_the_devil',
  DEBTRIDDEN = 'drawback:debtridden',
  DUNDERHEAD = 'drawback:dunderhead',
  EUNUCH = 'drawback:eunuch',
  LILYLIVERED = 'drawback:lilylivered',
  MELANCHOLIC_TEMPERAMENT = 'drawback:melancholic_temperament',
  NEER_DO_WELL = 'drawback:neer_do_well',
  NEMESIS = 'drawback:nemesis',
  PAINKILLER = 'drawback:painkiller',
  PERSECUTION_COMPLEX = 'drawback:persecution_complex',
  PHLEGMATIC_TEMPERAMENT = 'drawback:phlegmatic_temperament',
  SANGUINE_TEMPERAMENT = 'drawback:sanguine_temperament',
  SOUR_STOMACH = 'drawback:sour_stomach',
  SPLIT_FACE = 'drawback:split_face',
  VETERANS_BOOT = 'drawback:veterans_boot',
  VETERANS_EYE = 'drawback:veterans_eye',
  VETERANS_HAND = 'drawback:veterans_hand',
  VETERANS_LEG = 'drawback:veterans_leg',
  WEAK_LUNGS = 'drawback:weak_lungs',
  //
  LURE_OF_POWER = 'drawback:lure_of_power',
  CRUEL_TUTELAGE = 'drawback:cruel_tutelage',
  THE_END_IS_NIGH = 'drawback:the_end_is_nigh',
  WHEN_THE_DEALINS_DONE = 'drawback:when_the_dealins_done',
  CRESTFALLEN = 'drawback:crestfallen',
  ABSTAIN_FROM_VIOLENCE = 'drawback:abstain_from_violence',
  SALT_OF_THE_EARTH = 'drawback:salt_of_the_earth',
  NOTORIOUS_MOOCHER = 'drawback:notorious_moocher',
  WASHED_UP = 'drawback:washed_up',
  DEATH_WISH = 'drawback:death_wish',
  ALTER_EGO = 'drawback:alter_ego',
  WARP_DISSONANCE = 'drawback:warp_dissonance',
  ASSASSINS_CREED = 'drawback:assassins_creed',
  THE_CUSTODIANS_COMMANDMENTS = 'drawback:the_custodians_commandments',
  ENJOY_THE_SILENCE = 'drawback:enjoy_the_silence',
  // LURE_OF_POWER = 'drawback:aaaa',
  // LURE_OF_POWER = 'drawback:aaaa',
  // LURE_OF_POWER = 'drawback:aaaa',
  // LURE_OF_POWER = 'drawback:aaaa',
}